tcadserver: incorporate file descriptor table.
Before, tcad was a true key-value store, and each request specified the key.
This doesn't quite work with the driving use-casel, which is using tcad to
implement a shared memory, as would allow shared memory is more-or-less a
capability that is passed to child processes.

This breaks the testclient/tcadclient.c; will fix in the next commit.